// src/api.admin/users/dto/user.create.interface.ts
export interface IUserCreate {
    readonly lang_id: number;
    readonly parent_id: number;
    readonly referrer_id: number;
    readonly email: string;
    readonly password: string;
    readonly name: string;
    readonly wallet: string;
    readonly active: boolean;
    readonly money: number;
    readonly points: number;
    readonly paid_at: Date;
    readonly paid_until: Date;
    readonly children_limit: number;
    readonly freetasks: number;
    readonly freetasks_viewed_at: string;
    readonly referral_percent: number;
    readonly tg_id: number;
    readonly tg_active: boolean;
    readonly tg_tasks: boolean;
    readonly tg_guides: boolean;
    readonly tg_articles: boolean;
    readonly tg_deadlines: boolean;
    readonly tg_invite: string;
    readonly tz: number;
}